Tatsumi code location shuffle
